+++ README.md
@@ -106,6 +106,6 @@ To stabilize training, we used the Adam hyperparameters from [Kapturowski et al.
 
 Finally, note that the numbers for the MDP case are provisional, as we sample a stochastic reward at each timestep from a normal distribution 
 with unit variance, and resample the means when instantiating a new MDP. It is possible, and perhaps likely, that Duan et al., 2016 
-sampled the means once when creating the benchmark, and then sampled new deterministic reward for each MDP from a normal distribution 
-with these means and with unit variances. (If true, this could make their results a bit difficult to reproduce, since the distribution over MDPs 
+sampled the means once when creating the benchmark, and then sampled new deterministic rewards from a normal distribution 
+with these means and with unit variances, for each state-action pair in each newly sampled MDP. (If true, this could make their results a bit difficult to reproduce, since the distribution over MDPs 
 would depend heavily on the random means sampled from the hyperprior.)
